+++ toys-core/test/definition/test_tool.rb
@@ -595,8 +595,8 @@ describe Toys::Definition::Tool do
   describe "source info" do
     let(:source_path) { File.expand_path(__FILE__) }
     let(:source_path2) { File.expand_path(__dir__) }
-    let(:source_info) { Toys::Definition::SourceInfo.create_path_root(source_path) }
-    let(:source_info2) { Toys::Definition::SourceInfo.create_path_root(source_path2) }
+    let(:source_info) { Toys::SourceInfo.create_path_root(source_path) }
+    let(:source_info2) { Toys::SourceInfo.create_path_root(source_path2) }
 
     it "starts at nil" do
       assert_nil(tool.source_info)
@@ -703,7 +703,7 @@ describe Toys::Definition::Tool do
   describe "context directory" do
     let(:source_path) { File.expand_path(__FILE__) }
     let(:default_context_dir) { File.expand_path(__dir__) }
-    let(:source_info) { Toys::Definition::SourceInfo.create_path_root(source_path) }
+    let(:source_info) { Toys::SourceInfo.create_path_root(source_path) }
 
     it "defaults to nil" do
       assert_nil(tool.context_directory)